requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
Brotli==1.0.9
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
cryptography==36.0.1
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.3
Django==3.2.8
django-allauth==0.47.0
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-dbbackup==3.3.0
django-easy-pdf==0.1.1
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.6.1
django-mathfilters==1.0.0
django-xhtml2pdf==0.0.4
djangorestframework==3.13.1
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.3.0
Flask==2.0.3
fonttools==4.29.1
future==0.18.2
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
inflect==5.6.0
itsdangerous==2.1.0
Jinja2==3.0.3
Js2Py==0.71
libretranslatepy==2.1.1
lxml==4.7.1
MarkupPy==1.14
MarkupSafe==2.1.0
num2words==0.5.10
numpy==1.21.5
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
pdfkit==1.0.0
Pillow==8.4.0
pipwin==0.5.1
platformdirs==2.4.0
psycopg2==2.9.3
PyAudio @ file:///C:/Users/dell/pipwin/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycparser==2.21
pydyf==0.1.2
pyjsparser==2.7.1
PyJWT==2.3.0
pyparsing==3.0.6
PyPDF2==1.26.0
pyphen==0.12.0
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pyttsx3==2.90
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
PyYAML==6.0
reportlab==3.6.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tablib==3.1.0
tinycss2==1.1.1
translate==3.6.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
virtualenv==20.8.1
weasyprint==54.1
webencodings==0.5.1
Werkzeug==2.0.3
wikipedia==1.4.0
wkhtmltopdf==0.2
xhtml2pdf==0.2.5
xlrd==1.2.0
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.6.0
zopfli==0.1.9
